Error handling
git-svn-id: https://subversion.assembla.com/svn/cs6140@50 dcdcd5bf-5c71-4fd8-8603-51941cb4daa4

+++ Kruskal/src/Graph.h
@@ -15,7 +15,7 @@ struct GraphEdge
 {
 	struct GraphNode* startNode;
 	struct GraphNode* endNode;
-	int label;
+	double label;
 	struct GraphEdge* nextEdge;
 };
